+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Added support for Post Mutation parameters.
 - Added support for handling return values from server mutations.
 - Added support for custom transit handlers on the client side. Server side is coming in a release soon.
+- Added support for turning on/off Om path optimization
 
 0.5.7
 -----